# specs/001-add-auth-pages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Pages Flow API Contract
  version: 0.1.0
  description: Contract for frontend register/verify-code/login/logout flows.
servers:
  - url: http://localhost:3000
paths:
  /auth/register:
    post:
      summary: Register new account and enter pending verification state
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration accepted and verification code sent by email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/verify-code:
    post:
      summary: Verify account by email + registration code
      operationId: verifyCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Code verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyCodeResponse'
        '400':
          description: Invalid/expired/used code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success; session established.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Wrong credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account not verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      summary: Logout current session
      operationId: logout
      responses:
        '200':
          description: Logout success (idempotent).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'

components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    RegisterResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [PENDING_CODE]
        message:
          type: string

    VerifyCodeRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 4
          maxLength: 32

    VerifyCodeResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [VERIFIED]
        message:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      required: [status, message, user]
      properties:
        status:
          type: string
          enum: [AUTHENTICATED]
        message:
          type: string
        user:
          type: object
          required: [id, email]
          properties:
            id:
              type: string
            email:
              type: string
              format: email

    LogoutResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [ANONYMOUS]
        message:
          type: string

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: CODE_EXPIRED
        message:
          type: string
        action:
          type: string
          nullable: true